BE Module Role Permission

+++ app/Http/Controllers/Admin/AccountController.php
@@ -10,17 +10,45 @@ use App\Http\Requests\AccountUpdateRequest;
 
 class AccountController extends Controller
 {
+    /*
+    * Global Variable for Access Page
+    */
+    public $accessPage = [];
+
+    /*
+    * Get Access for Controller
+    */
+    public function get_access()
+    {
+        $this->accessPage = $this->get_access_per_page('Account');
+
+        $data = [
+            "Create" => (int) $this->accessPage['Create'],
+            "Read" => (int) $this->accessPage['Read'],
+            "Update" => (int) $this->accessPage['Update'],
+            "Delete" => (int) $this->accessPage['Delete'],
+        ];
+
+        return $data;
+    }
+
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        try {
-            $accounts = User::paginate(25);
-            return view('admin.setting.account.index',['accounts' => $accounts, 'access' => $this->get_access_per_page('Account')]);
-        } catch(\Illuminate\Database\QueryException $e) {
-            \Illuminate\Support\Facades\Log::error($e->getMessage());
-            return redirect()->back()->with('failed', $e->getMessage());
+        $access = $this->get_access();
+
+        if (!isset($access['Read']) || $access['Read'] != 1) {
+            return redirect()->back()->with('failed', "You don't have authority");
+        } else {
+            try {
+                $accounts = User::paginate(25);
+                return view('admin.setting.account.index', ['accounts' => $accounts, 'access' => $this->get_access_per_page('Account')]);
+            } catch (\Illuminate\Database\QueryException $e) {
+                \Illuminate\Support\Facades\Log::error($e->getMessage());
+                return redirect()->back()->with('failed', $e->getMessage());
+            }
         }
     }
 
@@ -29,11 +57,17 @@ class AccountController extends Controller
      */
     public function create()
     {
-        try {
-            return view('admin.setting.account.create');
-        } catch(\Illuminate\Database\QueryException $e) {
-            \Illuminate\Support\Facades\Log::error($e->getMessage());
-            return redirect()->back()->with('failed', $e->getMessage());
+        $access = $this->get_access();
+
+        if (!isset($access['Create']) || $access['Create'] != 1) {
+            return redirect()->back()->with('failed', "You don't have authority");
+        } else {
+            try {
+                return view('admin.setting.account.create');
+            } catch (\Illuminate\Database\QueryException $e) {
+                \Illuminate\Support\Facades\Log::error($e->getMessage());
+                return redirect()->back()->with('failed', $e->getMessage());
+            }
         }
     }
 
@@ -42,17 +76,23 @@ class AccountController extends Controller
      */
     public function store(AccountStoreRequest $request, AccountService $accountService)
     {
-        try {
-            $accountService->store(
-                $request->validated()
-            );
+        $access = $this->get_access();
 
-            return redirect()
-                ->route('setting.account.index')
-                ->with('success', 'Account created successfully.');
-        } catch(\Illuminate\Database\QueryException $e) {
-            \Illuminate\Support\Facades\Log::error($e->getMessage());
-            return redirect()->back()->with('failed', $e->getMessage());
+        if (!isset($access['Read']) || $access['Read'] != 1) {
+            return redirect()->back()->with('failed', "You don't have authority");
+        } else {
+            try {
+                $accountService->store(
+                    $request->validated()
+                );
+
+                return redirect()
+                    ->route('setting.account.index')
+                    ->with('success', 'Account created successfully.');
+            } catch (\Illuminate\Database\QueryException $e) {
+                \Illuminate\Support\Facades\Log::error($e->getMessage());
+                return redirect()->back()->with('failed', $e->getMessage());
+            }
         }
     }
 
@@ -61,11 +101,17 @@ class AccountController extends Controller
      */
     public function show(User $user)
     {
-        try {
-            //
-        } catch(\Illuminate\Database\QueryException $e) {
-            \Illuminate\Support\Facades\Log::error($e->getMessage());
-            return redirect()->back()->with('failed', $e->getMessage());
+        $access = $this->get_access();
+
+        if (!isset($access['Read']) || $access['Read'] != 1) {
+            return redirect()->back()->with('failed', "You don't have authority");
+        } else {
+            try {
+                //
+            } catch (\Illuminate\Database\QueryException $e) {
+                \Illuminate\Support\Facades\Log::error($e->getMessage());
+                return redirect()->back()->with('failed', $e->getMessage());
+            }
         }
     }
 
@@ -74,11 +120,17 @@ class AccountController extends Controller
      */
     public function edit(User $user)
     {
-        try {
-            return view('admin.setting.account.edit',['account' => $user->find(request()->segment(3))]);
-        } catch(\Illuminate\Database\QueryException $e) {
-            \Illuminate\Support\Facades\Log::error($e->getMessage());
-            return redirect()->back()->with('failed', $e->getMessage());
+        $access = $this->get_access();
+
+        if (!isset($access['Update']) || $access['Update'] != 1) {
+            return redirect()->back()->with('failed', "You don't have authority");
+        } else {
+            try {
+                return view('admin.setting.account.edit', ['account' => $user->find(request()->segment(3))]);
+            } catch (\Illuminate\Database\QueryException $e) {
+                \Illuminate\Support\Facades\Log::error($e->getMessage());
+                return redirect()->back()->with('failed', $e->getMessage());
+            }
         }
     }
 
@@ -87,18 +139,24 @@ class AccountController extends Controller
      */
     public function update(AccountUpdateRequest $request, User $user, AccountService $accountService)
     {
-        try {
-            $accountService->update(
-                $user->find(request()->segment(3)),
-                $request->validated()
-            );
+        $access = $this->get_access();
 
-            return redirect()
-                ->route('setting.account.index')
-                ->with('success', 'Account updated successfully.');
-        } catch(\Illuminate\Database\QueryException $e) {
-            \Illuminate\Support\Facades\Log::error($e->getMessage());
-            return redirect()->back()->with('failed', $e->getMessage());
+        if (!isset($access['Update']) || $access['Update'] != 1) {
+            return redirect()->back()->with('failed', "You don't have authority");
+        } else {
+            try {
+                $accountService->update(
+                    $user->find(request()->segment(3)),
+                    $request->validated()
+                );
+
+                return redirect()
+                    ->route('setting.account.index')
+                    ->with('success', 'Account updated successfully.');
+            } catch (\Illuminate\Database\QueryException $e) {
+                \Illuminate\Support\Facades\Log::error($e->getMessage());
+                return redirect()->back()->with('failed', $e->getMessage());
+            }
         }
     }
 
@@ -107,17 +165,23 @@ class AccountController extends Controller
      */
     public function destroy(User $user, AccountService $accountService)
     {
-        try {
-            $accountService->destroy(
-                $user->find(request()->segment(3))
-            );
+        $access = $this->get_access();
 
-            return redirect()
-                ->route('setting.account.index')
-                ->with('success', 'Account deleted successfully.');
-        } catch(\Illuminate\Database\QueryException $e) {
-            \Illuminate\Support\Facades\Log::error($e->getMessage());
-            return redirect()->back()->with('failed', $e->getMessage());
+        if (!isset($access['Delete']) || $access['Delete'] != 1) {
+            return redirect()->back()->with('failed', "You don't have authority");
+        } else {
+            try {
+                $accountService->destroy(
+                    $user->find(request()->segment(3))
+                );
+
+                return redirect()
+                    ->route('setting.account.index')
+                    ->with('success', 'Account deleted successfully.');
+            } catch (\Illuminate\Database\QueryException $e) {
+                \Illuminate\Support\Facades\Log::error($e->getMessage());
+                return redirect()->back()->with('failed', $e->getMessage());
+            }
         }
     }
 }
