+++ app/Http/Controllers/Admin/GroupController.php
@@ -10,19 +10,47 @@ use Illuminate\Support\Facades\Log;
 
 class GroupController extends Controller
 {
+    /*
+    * Global Variable for Access Page
+    */
+    public $accessPage = [];
+
+    /*
+    * Get Access for Controller
+    */
+    public function get_access()
+    {
+        $this->accessPage = $this->get_access_per_page('Roles');
+
+        $data = [
+            "Create" => (int) $this->accessPage['Create'],
+            "Read" => (int) $this->accessPage['Read'],
+            "Update" => (int) $this->accessPage['Update'],
+            "Delete" => (int) $this->accessPage['Delete'],
+        ];
+
+        return $data;
+    }
+
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        try {
-            return view('admin.setting.role.index', [
-                'roles' => Group::paginate(15),
-                'access' => $this->get_access_per_page('Roles')
-            ]);
-        } catch (\Illuminate\Database\QueryException $e) {
-            Log::error($e->getMessage());
-            return redirect()->back()->with('failed', 'Gagal memuat data role.');
+        $access = $this->get_access();
+
+        if (!isset($access['Read']) || $access['Read'] != 1) {
+            return redirect()->back()->with('failed', "You don't have authority");
+        } else {
+            try {
+                return view('admin.setting.role.index', [
+                    'roles' => Group::paginate(15),
+                    'access' => $this->get_access_per_page('Roles')
+                ]);
+            } catch (\Illuminate\Database\QueryException $e) {
+                Log::error($e->getMessage());
+                return redirect()->back()->with('failed', 'Gagal memuat data role.');
+            }
         }
     }
 
@@ -31,14 +59,20 @@ class GroupController extends Controller
      */
     public function create(RoleService $roleService)
     {
-        try {
-            return view('admin.setting.role.create', [
-                'page_distincts' => $roleService->getPageDistincts(),
-                'pages'          => \App\Models\Page::all(),
-            ]);
-        } catch (\Illuminate\Database\QueryException $e) {
-            Log::error($e->getMessage());
-            return redirect()->back()->with('failed', 'Gagal memuat form tambah role.');
+        $access = $this->get_access();
+
+        if (!isset($access['Create']) || $access['Create'] != 1) {
+            return redirect()->back()->with('failed', "You don't have authority");
+        } else {
+            try {
+                return view('admin.setting.role.create', [
+                    'page_distincts' => $roleService->getPageDistincts(),
+                    'pages'          => \App\Models\Page::all(),
+                ]);
+            } catch (\Illuminate\Database\QueryException $e) {
+                Log::error($e->getMessage());
+                return redirect()->back()->with('failed', 'Gagal memuat form tambah role.');
+            }
         }
     }
 
@@ -47,17 +81,23 @@ class GroupController extends Controller
      */
     public function store(Request $request, RoleService $roleService)
     {
-        $request->validate([
-            'group_name' => 'required|string|max:255',
-        ]);
+        $access = $this->get_access();
 
-        try {
-            $roleService->store($request->all());
-
-            return redirect()->route('setting.role.index')->with('success', 'Role berhasil ditambahkan!');
-        } catch (\Exception $e) {
-            Log::error($e->getMessage());
-            return redirect()->back()->with('failed', 'Gagal menyimpan data role.');
+        if (!isset($access['Create']) || $access['Create'] != 1) {
+            return redirect()->back()->with('failed', "You don't have authority");
+        } else {
+            $request->validate([
+                'group_name' => 'required|string|max:255',
+            ]);
+    
+            try {
+                $roleService->store($request->all());
+    
+                return redirect()->route('setting.role.index')->with('success', 'Role berhasil ditambahkan!');
+            } catch (\Exception $e) {
+                Log::error($e->getMessage());
+                return redirect()->back()->with('failed', 'Gagal menyimpan data role.');
+            }
         }
     }
 
@@ -66,11 +106,17 @@ class GroupController extends Controller
      */
     public function show(Group $group)
     {
-        try {
-            //
-        } catch (\Illuminate\Database\QueryException $e) {
-            Log::error($e->getMessage());
-            return redirect()->back()->with('failed', $e->getMessage());
+        $access = $this->get_access();
+
+        if (!isset($access['Read']) || $access['Read'] != 1) {
+            return redirect()->back()->with('failed', "You don't have authority");
+        } else {
+            try {
+                //
+            } catch (\Illuminate\Database\QueryException $e) {
+                \Illuminate\Support\Facades\Log::error($e->getMessage());
+                return redirect()->back()->with('failed', $e->getMessage());
+            }
         }
     }
 
@@ -79,18 +125,25 @@ class GroupController extends Controller
      */
     public function edit(Group $group, RoleService $roleService)
     {
-        try {
-            return view('admin.setting.role.edit', [
-                'page_distincts' => $roleService->getPageDistincts(),
-                'pages'          => \App\Models\GroupPage::leftJoin('pages', 'pages.id', '=', 'group_pages.page_id')
-                                        ->where('group_id', '=', $group->id)
-                                        ->get(),
-                'group'          => $group,
-            ]);
-        } catch (\Illuminate\Database\QueryException $e) {
-            Log::error($e->getMessage());
-            return redirect()->back()->with('failed', 'Gagal memuat data edit role.');
+        $access = $this->get_access();
+
+        if (!isset($access['Update']) || $access['Update'] != 1) {
+            return redirect()->back()->with('failed', "You don't have authority");
+        } else {
+            try {
+                return view('admin.setting.role.edit', [
+                    'page_distincts' => $roleService->getPageDistincts(),
+                    'pages'          => \App\Models\GroupPage::leftJoin('pages', 'pages.id', '=', 'group_pages.page_id')
+                                            ->where('group_id', '=', $group->id)
+                                            ->get(),
+                    'group'          => $group,
+                ]);
+            } catch (\Illuminate\Database\QueryException $e) {
+                Log::error($e->getMessage());
+                return redirect()->back()->with('failed', 'Gagal memuat data edit role.');
+            }
         }
+        
     }
 
     /**
@@ -98,18 +151,24 @@ class GroupController extends Controller
      */
     public function update(Request $request, Group $group, RoleService $roleService)
     {
-        // Validasi input sebelum proses update
-        $request->validate([
-            'group_name' => 'required|string|max:255',
-        ]);
+        $access = $this->get_access();
 
-        try {
-            $roleService->update($group, $request->all());
-
-            return redirect()->route('setting.role.index')->with('success', 'Role berhasil diperbarui!');
-        } catch (\Exception $e) {
-            Log::error($e->getMessage());
-            return redirect()->back()->with('failed', 'Gagal memperbarui data role.');
+        if (!isset($access['Update']) || $access['Update'] != 1) {
+            return redirect()->back()->with('failed', "You don't have authority");
+        } else {
+            // Validasi input sebelum proses update
+            $request->validate([
+                'group_name' => 'required|string|max:255',
+            ]);
+    
+            try {
+                $roleService->update($group, $request->all());
+    
+                return redirect()->route('setting.role.index')->with('success', 'Role berhasil diperbarui!');
+            } catch (\Exception $e) {
+                Log::error($e->getMessage());
+                return redirect()->back()->with('failed', 'Gagal memperbarui data role.');
+            }
         }
     }
 
@@ -118,13 +177,19 @@ class GroupController extends Controller
      */
     public function destroy(Group $group, RoleService $roleService)
     {
-        try {
-            $roleService->destroy($group);
+        $access = $this->get_access();
 
-            return redirect()->route('setting.role.index')->with('success', 'Role berhasil dihapus!');
-        } catch (\Exception $e) {
-            Log::error($e->getMessage());
-            return redirect()->back()->with('failed', 'Gagal menghapus data role.');
+        if (!isset($access['Delete']) || $access['Delete'] != 1) {
+            return redirect()->back()->with('failed', "You don't have authority");
+        } else {
+            try {
+                $roleService->destroy($group);
+    
+                return redirect()->route('setting.role.index')->with('success', 'Role berhasil dihapus!');
+            } catch (\Exception $e) {
+                Log::error($e->getMessage());
+                return redirect()->back()->with('failed', 'Gagal menghapus data role.');
+            }
         }
     }
 }
